requirements.txt
annotated-types==0.7.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
importlib_metadata==8.7.0
Markdown==3.9
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
pytz==2025.2
requests==2.32.5
rsa==4.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
zipp==3.23.0
